+++ apps/client/src/app/modules/list/data/extractor/voyages-extractor.ts
@@ -4,12 +4,13 @@ import { ItemData } from '../../../../model/garland-tools/item-data';
 import { DataType } from '../data-type';
 import { Item } from '../../../../model/garland-tools/item';
 import { GarlandToolsService } from '../../../../core/api/garland-tools.service';
-import { airshipVoyages } from '../../../../core/data/sources/airship-voyages';
-import { submarineVoyages } from '../../../../core/data/sources/submarine-voyages';
+import { LazyDataService } from 'apps/client/src/app/core/data/lazy-data.service';
+import { ExplorationType } from '../../../../model/other/exploration-type';
+import { uniqBy } from 'lodash';
 
 export class VoyagesExtractor extends AbstractExtractor<I18nName[]> {
 
-  constructor(gt: GarlandToolsService) {
+  constructor(gt: GarlandToolsService, private lazyData: LazyDataService) {
     super(gt);
   }
 
@@ -22,14 +23,17 @@ export class VoyagesExtractor extends AbstractExtractor<I18nName[]> {
   }
 
   protected canExtract(item: Item): boolean {
-    return item.voyages !== undefined;
+    return item.voyages !== undefined || this.lazyData.data.voyageSources[item.id] !== undefined;
   }
 
   protected doExtract(item: Item, itemData: ItemData): I18nName[] {
     const voyages: I18nName[] = [];
     if (item.voyages !== undefined) {
       item.voyages.forEach(v => {
-        const entry = [...airshipVoyages, ...submarineVoyages].find(e => e.en.toLowerCase() === v.toLowerCase());
+        const entry = [
+          ...Object.values<I18nName>(this.lazyData.data.airshipVoyages),
+          ...Object.values<I18nName>(this.lazyData.data.submarineVoyages)
+        ].find(e => e.en.toLowerCase() === v.toLowerCase());
         if (entry) {
           voyages.push({ ...entry });
         } else {
@@ -42,7 +46,12 @@ export class VoyagesExtractor extends AbstractExtractor<I18nName[]> {
         }
       });
     }
-    return voyages;
+    if (this.lazyData.data.voyageSources[item.id] !== undefined) {
+      this.lazyData.data.voyageSources[item.id].forEach(({ type, id }) => {
+        voyages.push((type === ExplorationType.AIRSHIP ? this.lazyData.data.airshipVoyages : this.lazyData.data.submarineVoyages)[id]);
+      });
+    }
+    return uniqBy(voyages, 'en');
   }
 
 }